Style-profiling/mlp-data-service/pyproject.toml
[tool.poetry]
name = "mlp-data-service"
version = "0.8.1"
description = ""
authors = ["AI-선행기술개발팀 <ai-dev@diquest.com>"]

[tool.poetry.dependencies]
python = "~3.8"
Django = "3.1.13"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.11.0"
djongo = "^1.3.6"
pymongo = "^3.12.1"
tensorflow = "^2.10.0"
tensorboard = "^2.10.1"
django-background-tasks = "^1.2.5"
Pillow = "^9.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
